Switch to using triple buffering
